+++ src/pages/FormPage.tsx
@@ -1,62 +1,38 @@
-/* eslint-disable prettier/prettier */
-import React, { useState } from 'react';
-import { useForm } from 'react-hook-form';
-import Cards from '../components/Cards';
-import { Checkbox } from '../components/Form/Checkbox';
-import { DateInput } from '../components/Form/DateInput';
-import { FileInput } from '../components/Form/FileInput';
-import { GenderInput } from '../components/Form/GenderInput';
-import { NameInput } from '../components/Form/NameInput';
-import { SpeciesInput } from '../components/Form/SpeciesInput';
-import { StatusInput } from '../components/Form/StatusInput';
-import { FormInputs, ICard } from '../types/types';
+import React, { useEffect, useState } from 'react';
+import CardModal from '../components/Card/CardModal';
+import Cards from '../components/Card/Cards';
+import { FormComponent } from '../components/Form/FormComponent';
+import { ICard } from '../types/types';
 
 export function FormPage() {
-  const [cards, setCards] = useState<ICard[]>([])
-  const [modal, setModal] = useState(false);
-  const {
-    register,
-    handleSubmit,
-    reset,
-    formState: { errors },
-  } = useForm<FormInputs>({ mode: 'onSubmit', reValidateMode: 'onSubmit' });
+  const [addModal, setAddModal] = useState(false);
+  const [modal, setModal] = useState(0);
+  const [cards, setCards] = useState<ICard[]>([]);
+  const [currentCard, setCurrentCard] = useState<ICard>({});
 
-  const onSubmit = handleSubmit((data) => {
-    const newCard = {
-      id: cards.length,
-      image: URL.createObjectURL((data.FileInput[0]) as unknown as MediaSource),
-      name: data.NameInput,
-      species: data.SpeciesInput,
-      status: data.StatusInput,
-      gender: data.GenderInput,
-      created: data.DateInput,
+  useEffect(() => {
+    console.log('currentCard', currentCard);
+    if (Object.keys(currentCard).length !== 0) {
+      setCards([...cards, currentCard]);
+      setCurrentCard({});
+      setAddModal(true);
+      setTimeout(() => {
+        setAddModal(false);
+      }, 1500);
     }
-    const cardsArr = cards;
-    cardsArr.push(newCard)
-    setCards(cardsArr)
-    reset();
-    setModal(true);
-  });
+  }, [cards, currentCard]);
 
   return (
     <main className="mainForm" data-testid="FormPage">
-      <form className={'form'} onSubmit={onSubmit} >
-        <NameInput register={register} errors={errors} />
-        <SpeciesInput register={register} errors={errors} />
-        <GenderInput register={register} errors={errors} />
-        <StatusInput register={register} errors={errors} />
-        <DateInput register={register} errors={errors} />
-        <FileInput register={register} errors={errors} />
-        <Checkbox register={register} errors={errors} />
-        <input type="submit" />
-      </form>
+      {!(modal === 0) && <CardModal id={modal} cards={cards} setModal={setModal} />}
+      <FormComponent setCurrentCard={setCurrentCard} />
       <>
-        {cards.length > 0 ? <Cards cards={[...cards]} /> : ''}
-        {modal ? (
+        {cards.length > 0 ? <Cards cards={[...cards]} setModal={setModal} /> : ''}
+        {addModal ? (
           <div className="modal">
             <h2>Confirmation message:</h2>
             <p>Data has been saved!</p>
-            <button onClick={() => setModal(false)}>ok!</button>
+            <button onClick={() => setAddModal(false)}>ok!</button>
           </div>
         ) : (
           ''
